requirement.txt
fastapi
uvicorn
streamlit
google-api-python-client
google-auth
google-auth-oauthlib
python-dateutil
requests
